## Why
依存クレートのバージョンアップ（rand 0.10 等）に伴うビルドエラーと、
新しいClippy（rustc 1.93.0相当）が検出した警告を解消する必要があった。

## What
- Cargo.toml/Cargo.lock: rand を 0.10 へアップグレード（bitflags 等も更新）
- leptonica-region: rand 0.10 API 変更対応（thread_rng→rng, gen→random, Rng→RngExt）
- workspace 全体: clippy 警告 49件を修正
  - needless_range_loop: iter().enumerate() / iter() に変換
  - manual_div_ceil / manual_is_multiple_of / manual_range_contains: 標準メソッドに置換
  - collapsible_if: let チェーンに統合
  - unnecessary_cast / len_zero / approx_constant 等: 各種イディオム修正

## Impact
- leptonica-core, leptonica-region, leptonica-color, leptonica-filter の各クレート
- 機能変更なし、ビルド・テスト・lint がクリーンな状態に復元

PR: #114

+++ crates/leptonica-color/tests/quantize_ext_reg.rs
@@ -217,7 +217,7 @@ fn test_quant_from_cmap_gray() {
     let pix = Pix::new(100, 1, PixelDepth::Bit8).unwrap();
     let mut pm = pix.try_into_mut().unwrap();
     for x in 0..100u32 {
-        pm.set_pixel_unchecked(x, 0, (x * 255 / 99) as u32);
+        pm.set_pixel_unchecked(x, 0, x * 255 / 99);
     }
     let pix: Pix = pm.into();
     let mut cmap = PixColormap::new(8).unwrap();
