+++ crates/leptonica-filter/tests/adaptmap_morph_reg.rs
@@ -17,7 +17,7 @@ fn make_gray_test_image() -> Pix {
     for y in 0..80 {
         for x in 0..80 {
             // Background gradient (brighter right side)
-            let bg = 120 + (x as u32 * 2).min(135);
+            let bg = 120 + (x * 2).min(135);
             // Dark "text" region in center
             let val = if x > 20 && x < 60 && y > 20 && y < 60 {
                 bg / 3
